refactor(terraform-mirror): extract card, version row, chips and tool constants
First step of splitting TerraformMirrorPage.tsx (1402 lines) along its own
seams. These four pieces are pure presentation/vocabulary shared by more than
one dialog, so they move out before the dialog flows themselves:

- constants.ts        tool list, default upstream URLs, platform options and
                      the blank create form — one place to add a new tool
- StatusChips.tsx     the sync-status and tool chips used by the card, the
                      versions table and the history table
- MirrorConfigCard    one config's summary tile plus its action buttons
- MirrorVersionRow    one versions-table row plus its lazy platform fetch

Markup, props and the lazy-fetch behaviour are moved verbatim. All 43 existing
page tests still pass unchanged.
